myapp/myproject/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
pytest-django==4.7.0  # Add this line
